Moved Crawdad.dll to SwaMe and incorporated it into SwaMe
Crawdad is now the chief peakpicking method of SwaMe. PeakSymmetry is currently set to 1 and still needs to be added with FW at 5% which will be done in a subsequent commit.
Also wavelettransform moved to its own project. It is still currently used for iRT peptide search.
