+++ Console/Yamato.Console.csproj
@@ -9,10 +9,10 @@
 
   <ItemGroup>
     <PackageReference Include="CenterSpace.NMath.Standard.Windows.X64" Version="7.0.0.47" />
-    <PackageReference Include="CommandLineParser" Version="2.4.3" />
-    <PackageReference Include="MathNet.Numerics" Version="4.7.0" />
+    <PackageReference Include="CommandLineParser" Version="2.6.0" />
+    <PackageReference Include="MathNet.Numerics" Version="4.8.1" />
     <PackageReference Include="NETStandard.Library" Version="2.0.3" />
-    <PackageReference Include="NLog" Version="4.6.3" />
+    <PackageReference Include="NLog" Version="4.6.7" />
   </ItemGroup>
 
   <ItemGroup>
