+++ CrawDad/Crawdad.dll.metagen
@@ -1,0 +1,78 @@
+ImageRuntimeVersion: v4.0.30319
+Assembly Crawdad, Version=1.0.*, Culture=Invariant Language (Invariant Country): 
+	hash=SHA1, flags=PublicKey
+Assembly mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089: 
+	hash=None, flags=None
+Assembly System.Data, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089: 
+	hash=None, flags=None
+Assembly System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089: 
+	hash=None, flags=None
+Assembly System.Xml, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089: 
+	hash=None, flags=None
+Class crawdadyamatowrapper.paul.CrawdadPeakFinder: AutoLayout, AnsiClass, Class, Public, BeforeFieldInit
+    Void .ctor(): PrivateScope, Public, HideBySig, SpecialName, RTSpecialName
+  Methods:
+    HelloWorld(): PrivateScope, Public, HideBySig
+Class pwiz.Crawdad.CrawdadPeak: AutoLayout, AnsiClass, Class, Public, BeforeFieldInit
+    Void .ctor(crawpeaks.SlimCrawPeak*): PrivateScope, Public, HideBySig, SpecialName, RTSpecialName
+  Interfaces:
+    System.IDisposable
+  Properties:
+    Boolean Identified 'get ' : 
+    Boolean FwhmDegenerate 'get ' : 
+    Single Fwhm 'get ' : 
+    Single Height 'get ' : 
+    Single BackgroundArea 'get ' : 
+    Single Area 'get ' : 
+    Int32 Center 'get ' : 
+    Int32 Length 'get ' : 
+    Int32 EndIndex 'get ' : 
+    Int32 StartIndex 'get ' : 
+    Int32 TimeIndex 'get ' : 
+  Methods:
+    get_TimeIndex(): PrivateScope, Public, HideBySig, SpecialName
+    get_StartIndex(): PrivateScope, Public, HideBySig, SpecialName
+    get_EndIndex(): PrivateScope, Public, HideBySig, SpecialName
+    get_Length(): PrivateScope, Public, HideBySig, SpecialName
+    get_Center(): PrivateScope, Public, HideBySig, SpecialName
+    get_Area(): PrivateScope, Public, HideBySig, SpecialName
+    get_BackgroundArea(): PrivateScope, Public, HideBySig, SpecialName
+    get_Height(): PrivateScope, Public, HideBySig, SpecialName
+    get_Fwhm(): PrivateScope, Public, HideBySig, SpecialName
+    get_FwhmDegenerate(): PrivateScope, Public, HideBySig, SpecialName
+    get_Identified(): PrivateScope, Public, HideBySig, SpecialName
+    ResetBoundaries(Int32, Int32): PrivateScope, Public, HideBySig
+    SetIdentified(Int32[]): PrivateScope, Public, HideBySig
+    ToString(): PrivateScope, Public, Virtual, HideBySig
+    Dispose(): PrivateScope, Public, Final, Virtual, HideBySig
+Class pwiz.Crawdad.CrawdadPeakFinder: AutoLayout, AnsiClass, Class, Public, BeforeFieldInit
+    Void .ctor(): PrivateScope, Public, HideBySig, SpecialName, RTSpecialName
+  Interfaces:
+    System.IDisposable
+  Properties:
+    System.Collections.Generic.List`1[System.Single] Intensities1d 'get ' : 
+    System.Collections.Generic.List`1[System.Single] Intensities2d 'get ' : 
+    Single StdDev 'get set' : 
+    Single FullWidthHalfMax 'get set' : 
+    Boolean IsHeightAsArea 'get ' : 
+  Methods:
+    HelloWorld(basic_string<char\,std::char_traits<char>\,std::allocator<char> >*): PrivateScope, Public, HideBySig
+    SetChromatogram(IList`1, IList`1): PrivateScope, Public, HideBySig
+    SetChromatogram(IList`1, IList`1): PrivateScope, Public, HideBySig
+    get_IsHeightAsArea(): PrivateScope, Public, HideBySig, SpecialName
+    get_FullWidthHalfMax(): PrivateScope, Public, HideBySig, SpecialName
+    set_FullWidthHalfMax(Single): PrivateScope, Public, HideBySig, SpecialName
+    get_StdDev(): PrivateScope, Public, HideBySig, SpecialName
+    set_StdDev(Single): PrivateScope, Public, HideBySig, SpecialName
+    get_Intensities2d(): PrivateScope, Public, HideBySig, SpecialName
+    get_Intensities1d(): PrivateScope, Public, HideBySig, SpecialName
+    CalcPeaks(Int32, Int32[]): PrivateScope, Public, HideBySig
+    CalcPeaks(Int32): PrivateScope, Public, HideBySig
+    CalcPeaks(): PrivateScope, Public, HideBySig
+    GetPeak(Int32, Int32): PrivateScope, Public, HideBySig
+    Dispose(): PrivateScope, Public, Final, Virtual, HideBySig
+Class pwiz.Crawdad.CrtDebugHeap: AutoLayout, AnsiClass, Class, Public, BeforeFieldInit
+    Void .ctor(): PrivateScope, Public, HideBySig, SpecialName, RTSpecialName
+  Methods:
+    Checkpoint(): PrivateScope, Public, Static, HideBySig
+    DumpLeaks(Boolean(HasFieldMarshal)): PrivateScope, Public, Static, HideBySig
