+++ Yamato.sln
@@ -15,7 +15,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LibraryParser", "LibraryPar
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MzqcGenerator", "MzqcGenerator\MzqcGenerator.csproj", "{60EF5F87-175C-4BD9-AD1C-04E57A71232F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IRTSearcher", "IRTSearch\IRTSearcher.csproj", "{85419632-0C73-4AFB-93B7-F2556E422879}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IRTSearcher", "IRTSearch\IRTSearcher.csproj", "{85419632-0C73-4AFB-93B7-F2556E422879}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CWT", "CWT\CWT.csproj", "{95473CB6-E1A1-4E5A-96E4-D33E5332B037}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -51,6 +53,10 @@ Global
 		{85419632-0C73-4AFB-93B7-F2556E422879}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{85419632-0C73-4AFB-93B7-F2556E422879}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{85419632-0C73-4AFB-93B7-F2556E422879}.Release|Any CPU.Build.0 = Release|Any CPU
+		{95473CB6-E1A1-4E5A-96E4-D33E5332B037}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{95473CB6-E1A1-4E5A-96E4-D33E5332B037}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{95473CB6-E1A1-4E5A-96E4-D33E5332B037}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{95473CB6-E1A1-4E5A-96E4-D33E5332B037}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
